# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "An example of gin",
        "title": "Golang Gin API",
        "termsOfService": "运行地址：http://localhost/swagger/index.html",
        "contact": {},
        "license": {
            "name": "MIT //localhost:80"
        },
        "version": "2.0"
    },
    "paths": {
...
        "/api/books": {
            "get": {
                "description": "书籍列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "书籍"
                ],
                "summary": "获取书籍列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "title",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "pageNum",
                        "name": "pageNum",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "pageSize",
                        "name": "pageSize",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Book"
                        }
                    },
                    "400": {
                        "description": "{ \"code\": 400, \"message\": \"请求失败\" }",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "创建书籍",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "书籍"
                ],
                "summary": "创建书籍列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "title",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "year",
                        "name": "year",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "author",
                        "name": "actor",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "desc",
                        "name": "desc",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Book"
                        }
                    },
                    "400": {
                        "description": "{ \"code\": 400, \"message\": \"请求失败\" }",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/books/{id}": {
            "put": {
                "description": "书籍列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "书籍"
                ],
                "summary": "更新书籍列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Book"
                        }
                    },
                    "400": {
                        "description": "{ \"code\": 400, \"message\": \"id必传\" }",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "书籍列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "书籍"
                ],
                "summary": "删除书籍列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Book"
                        }
                    },
                    "400": {
                        "description": "{ \"code\": 400, \"message\": \"id必传\" }",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/films": {
            "get": {
                "description": "电影列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "电影"
                ],
                "summary": "获取电影列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "name",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "pageNum",
                        "name": "pageNum",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "pageSize",
                        "name": "pageSize",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Film"
                        }
                    },
                    "400": {
                        "description": "{ \"code\": 400, \"message\": \"请求失败\" }",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "创建电影",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "电影"
                ],
                "summary": "创建电影列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "name",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "year",
                        "name": "year",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "address",
                        "name": "address",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "actor",
                        "name": "actor",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "desc",
                        "name": "desc",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Film"
                        }
                    },
                    "400": {
                        "description": "{ \"code\": 400, \"message\": \"请求失败\" }",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/films/{id}": {
            "put": {
                "description": "电影列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "电影"
                ],
                "summary": "更新电影列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Film"
                        }
                    },
                    "400": {
                        "description": "{ \"code\": 400, \"message\": \"id必传\" }",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "电影列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "电影"
                ],
                "summary": "删除电影列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Film"
                        }
                    },
                    "400": {
                        "description": "{ \"code\": 400, \"message\": \"id必传\" }",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.Book": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "string",
                    "example": "作者"
                },
                "desc": {
                    "type": "string",
                    "example": "描述"
                },
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string",
                    "example": "书籍名称"
                },
                "year": {
                    "type": "string",
                    "example": "年份"
                }
            }
        },
        "model.Film": {
            "type": "object",
            "properties": {
                "actor": {
                    "type": "string",
                    "example": "演员"
                },
                "address": {
                    "type": "string",
                    "example": "出品地区"
                },
                "desc": {
                    "type": "string",
                    "example": "描述"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string",
                    "example": "电影名称"
                },
                "year": {
                    "type": "string",
                    "example": "年份"
                }
            }
        },
...
    }
}